requirements.txt
alembic==1.0.7
aniso8601==3.0.2
Click==7.0
Flask==1.0.2
Flask-GraphQL==2.0.0
Flask-JWT-Extended==3.17.0
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
graphene==2.1.3
graphene-sqlalchemy==2.1.0
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
passlib==1.7.1
promise==2.2.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
gunicorn
